weight and height converted to kilogram and meter

+++ app/src/main/java/com/ozan/cleanpokedex/ui/uimodel/pokemon/PhysicalInfoUiModel.kt
@@ -1,6 +1,6 @@
 package com.ozan.cleanpokedex.ui.uimodel.pokemon
 
 data class PhysicalInfoUiModel(
-    val value: Int,
+    val value: Float,
     val unit: String
 )
